% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_git_snapshot.R
\name{tar_git_snapshot}
\alias{tar_git_snapshot}
\title{Snapshot a data store Git repository.}
\usage{
tar_git_snapshot(
  store = targets::tar_config_get("store"),
  prompt = interactive(),
  verbose = TRUE
)
}
\arguments{
\item{store}{Character of length 1, path to the data store of the pipeline.
If \code{NULL}, the \code{store} setting is left unchanged in the
YAML configuration file (default: \verb{_targets.yaml}).
Usually, the data store lives at \verb{_targets}.
Set \code{store} to a custom directory
to specify a path other than \verb{_targets/}. The path need not exist
before the pipeline begins, and it need not end with "_targets",
but it must be writeable.
For optimal performance, choose a storage location
with fast read/write access.
If the argument \code{NULL}, the setting is not modified.
Use \code{\link[targets:tar_config_unset]{tar_config_unset()}} to delete a setting.}
}
\description{
Snapshot a Git repository for a \code{targets} data store.
}
\details{
\code{tar_git_snapshot()} does the following:
1.
}
\examples{
if (Sys.getenv("GITTARGETS_EXAMPLES") == "true") {
targets::tar_dir({ # Containing code does not modify the user's filespace.
targets::tar_script()
targets::tar_make()
list.files("_targets", all.files = TRUE)
tar_git_init()
list.files("_targets", all.files = TRUE)
})
}
}
\seealso{
Other git: 
\code{\link{tar_git_init}()},
\code{\link{tar_git_status}()}
}
\concept{git}
