% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_manifest.R
\name{zzz_write_manifest}
\alias{zzz_write_manifest}
\title{Write a manifest file.}
\usage{
zzz_write_manifest()
}
\value{
\code{NULL} (invisibly).
}
\description{
Not a user-side function.
}
\details{
Writes a \code{manifest.json} file so the \code{pkgdown} site
can be deployed to RStudio Connect as Git-backed content.
}
\examples{
zzz_write_manifest()
}
